Create SaSpt03
Adding a program for the standardized SaSpt03 macro that calculates supply air setpoints based on HtgLoad & ClgLoad.  This is usually used in conjunction with a SaFanSpdXX macro that also uses HtgLoad and ClgLoad to calculate supply fan speed.  The HtgLoad and ClgLoad are broken up into two bands.  The lower bands are used for calculating the setpoint and the upper bands are used to calculate the supply fan speed.